requirements.txt
python-docx==0.8.11
fpdf==1.7.2
Flask==2.3.3
gunicorn==21.2.0
